Encapsulated diagnostics behind a swappable no-op sink with a -d switch
* Split static OspreyDiagnostics into a call-site facade + OspreyFileDiagnostics sink (null = no-op default)
* Production carries no diagnostic sink; env vars self-enable it, -d forces the structured dump bundle
* Added -d / --diagnostics CLI flag wired through OspreyConfig + OspreyDiagnostics.Initialize
* Dump bodies unchanged (byte-stable); verified build + 372 tests + ReSharper 0-warning gate

See ai/todos/active/TODO-20260606_ospreysharp_diagnostics_di.md
